+++ CHANGELOG.md
@@ -12,6 +12,7 @@ _Release Date: ???_
 - `/entity/resources` merged into `/state/entity/details` endpoint
 - new paginable endpoints `/state/entity/page/{x}` for metadata, fungibles, fungible-vaults, non-fungibles, non-fungible-vaults and non-fungible-vault/ids. Cursors returned from `state/entity/details` can be later used to query for next pages in specific endpoints.
 - `/transaction/committed-details` endpoint operates on `intent_hash_hex` only now
+- new `/sbor/scrypto-encode` `/sbor/scrypto-decode` endpoints to encode and decode sbor.
 
 ### 0.1.1 - Betanet V2
 _Release Date: 14.02.2022_